backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.12.0
boto3==1.34.19
stripe==7.10.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.26.0
tenacity==8.2.3
ffmpeg-python==0.2.0
pydub==0.25.1
